rename config file
