+++ src/documentConfig.ts
